# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Medication:
    properties:
      activeIngredient:
        type: string
      companyHoldingRegistration:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      registrationNumber:
        type: string
      regulatoryCategory:
        type: string
      therapeuticClass:
        type: string
      updatedAt:
        type: string
    required:
    - activeIngredient
    - companyHoldingRegistration
    - name
    - registrationNumber
    - regulatoryCategory
    - therapeuticClass
    type: object
...
  requests.MedicationRequest:
    properties:
      CompanyHoldingRegistration:
        type: string
      activeIngredient:
        type: string
      name:
        type: string
      registrationNumber:
        type: string
      regulatoryCategory:
        type: string
      therapeuticClass:
        type: string
    required:
    - CompanyHoldingRegistration
    - activeIngredient
    - name
    - registrationNumber
    - regulatoryCategory
    - therapeuticClass
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: The objective this application is save history medicaments of person.
  title: Health History API
  version: "1.0"
paths:
...
  /medications:
    get:
      description: Return medication list
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            items:
              $ref: '#/definitions/models.Medication'
            type: array
      summary: List medications
      tags:
      - medications
    post:
      consumes:
      - application/json
      description: Create a new medication
      parameters:
      - description: Data of medication
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.MedicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successful response
          schema:
            $ref: '#/definitions/models.Medication'
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create medication
      tags:
      - medications
  /medications/:id:
    get:
      description: Return medication by id
      parameters:
      - description: Medication ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/models.Medication'
        "404":
          description: Medication not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Return medication
      tags:
      - medications
    put:
      consumes:
      - application/json
      description: Update a exist medication
      parameters:
      - description: Data of medication
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.MedicationRequest'
      - description: Medication ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/models.Medication'
        "400":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update  medication
      tags:
      - medications
...
swagger: "2.0"